doc(http-services): fix link text not rendering correctly
The [HTTP services docs](http://aurelia.io/hub.html#/doc/article/aurelia/fetch-client/latest/http-services/1) have a link to a page that shows which browsers implement the fetch standard. The text for that link contains the word "haven't", but this is not rendered correctly, as the text is truncated to "haven". This PR replaces "haven't" with "have not".